Refactor Loop device status and clear predictions when switching from one loop with predictions to one without
